Back component, cartState on the handleDelete, name on the cart
