+++ src/components/footer/Footer.module.css
@@ -26,3 +26,7 @@
     color: #fff;
     padding: auto;
 }
+.contactInfo{
+    display: flex;
+    padding: 0 1rem;
+}
